# llama-bench : use random tokens to improve accuracy with mixtral